backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
openpyxl==3.1.5
openai==1.58.1
python-dotenv==1.0.1
